CMakeExpandImportedTargets: Document as deprecated
The module is not needed anymore for try_compile or try_run.  It cannot
be used with CMP0022 NEW behavior due to generator expressions in
INTERFACE_LINK_LIBRARIES, so document it as deprecated.  Whatever
problems other than try_compile and try_run anyone tried to solve with
this module will have to be addressed another way.