FindHg: Update rejection of cygwin hg on Windows
The cygwin hg client is a text file with a '#!/bin/python" line.  This
cannot run on Windows.  Since KWSys commit faff2ab0 (Process: Wait for
children to terminate on Ctrl+C, 2015-06-30) the execute_process result
value is different in this case.  Update our check to account for it,
at least until that is addressed.