HP-UX: Fix shared library search flags on HP-UX.
In the HP-UX platform file, -Wl,-a-default is used for shared library
searches.  According to the reference manual for the HP compilers, the
default option will search for shared libraries and fallback to static
if not available.  This restricts the search to use shared libs
exclusively with the -Wl,-a,shared flag to put behavior in sync with all
other platforms.

See "HP Itanium-based Systems HP aC++/HP C Programmer's Guide"
Chapter "Tools and Libraries", Section "Creating and Using Libraries"