Utilities/Release: Install docs in unversioned dir on OS X (#15692)
The CMake.app directory will never overlap with another version of CMake
so we do not need to have a version number on the documentation
directory.  Configure our release binaries on OS X to place the
documentation in an unversioned directory under CMake.app so that
one can always use a path like

 /Applications/CMake.app/Contents/doc/cmake/html/index.html

to visit the documentation even after updating CMake.