variables: remove _LIB_DEPENDS variables
Introduces policy CMP0066 to avoid setting _LIB_DEPENDS cache variables
to track the dependencies of a target. Instead, a global property is
used to store the information since it is only required during a single
configuration anyways.

Benefits include reducing the size of the cache file, fewer variables to
look through in the cache, and INTERFACE libraries can now have the same
name as a target that previously existed (which errored out previously
because old _LIB_DEPENDS variables would cause an error when the
INTERFACE library was declared).