Fix autoexport feature for 64 bit Windows builds by not removing leading _
With 64 bit windows builds, there is no need to remove the leading
underscore from all the symbols.  This is because it does not have
one in the .obj file unless it is really in the name.  This did
not cause any trouble until VS 2015 which has some system functions
that have a leading underscore that end up in the .def file.