Cray: Do not support shared libraries if compiler links statically
When using the Cray programming environment on Linux the compiler
wrappers may force static linking.  In this case tell CMake that the
target architecture does not support shared libraries.  This will
prevent projects from accidentally adding shared libraries when the
toolchain does not support them.  It also helps CMake avoid linking with
flags that require shared libraries to be supported.