Server: Add protocol for querying content in context
The protocol currently requires the client to sent the entire editor
content as a JSON value.  The protocol should be extended to accept a
filepath where the editor content is stored.  Editors like vim keep such
a file.

Also to be considered is extending the protocol to allow editors to
specify only changed lines.