Error out if --target is specified multiple times
So far it was not clearly documented that --target is only supported to be
specified once. Even worse, specifying it multiple times would silently
ignore any previously specified targets and only build the last target.

Improve the docs in this regard and error out if --target is specified
more than once in order to prevent user errors.