cmake-mode.el: Fix help completion item lists with CMake >= 3.0 (#16019)
We run `cmake --help-*-list` to get a list of items for completion.
Since CMake < 3.0 always printed "cmake version ..." on the first line
of the output we have previously ignored the first line.  However, CMake
3.0 and above do not print the version line so we should not ignore the
first line or we miss one item.

Ideally we should filter the first line out if it is "cmake version ..."
in order to support CMake < 3.0 cleanly, but at worst the version line
will show up as a completion option so simply including the first line
is good enough for now.